Move sample Application class under 'server' package.
