+++ core-java-sample/src/main/java/org/spine3/sample/package-info.java
@@ -18,3 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+@ParametersAreNonnullByDefault
+package org.spine3.sample;
+
+import javax.annotation.ParametersAreNonnullByDefault;
