+++ core-java-sample/src/main/java/org/spine3/sample/server/EventLogger.java
@@ -17,7 +17,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.spine3.sample;
+package org.spine3.sample.server;
 
 import com.google.protobuf.Message;
 import org.slf4j.Logger;
